+++ src/main/java/com/braintreegateway/PaymentMethodNonceDetailsPayerInfo.java
@@ -9,8 +9,9 @@ public class PaymentMethodNonceDetailsPayerInfo {
     private String lastName;
     private String payerId;
     private String countryCode;
+    private Address billingAddress;
+
     // TODO coerce the address info to match the address object?
-    // private Address billingAddress;
     // private Address shippingAddress;
 
     public PaymentMethodNonceDetailsPayerInfo(NodeWrapper node) {
@@ -19,6 +20,12 @@ public class PaymentMethodNonceDetailsPayerInfo {
         lastName = node.findString("last-name");
         payerId = node.findString("payer-id");
         countryCode = node.findString("country-code");
+
+        NodeWrapper billingAddressNode = node.findFirst("billing-address");
+
+        if (billingAddressNode != null && !billingAddressNode.isBlank()) {
+            billingAddress = new Address(billingAddressNode);
+        }
     }
 
     public PaymentMethodNonceDetailsPayerInfo(Map<String, String> map) {
@@ -47,5 +54,33 @@ public class PaymentMethodNonceDetailsPayerInfo {
 
     public String getCountryCode() {
         return countryCode;
+    }
+
+    public Address getBillingAddress() {
+        return billingAddress;
+    }
+
+    public static class Address {
+        public String line1;
+        public String city;
+        public String state;
+        public String countryCode;
+        public String postalCode;
+
+        public Address(NodeWrapper node) {
+            line1 = node.findString("line1");
+            city = node.findString("city");
+            state = node.findString("state");
+            countryCode = node.findString("country-code");
+            postalCode = node.findString("postal-code");
+        }
+
+        public Address(Map<String, String> map) {
+            line1 = map.get("line1");
+            city = map.get("city");
+            state = map.get("state");
+            countryCode = map.get("country-code");
+            postalCode = map.get("postal-code");
+        }
     }
 }