+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fix null pointer exception when 3DS lookup returns a `422`
 * Add ThreeDSecure test payment method nonces
 * Add test `AuthenticationId`s
+* Add `DISPUTE_ACCEPTED`, `DISPUTE_DISPUTED`, and `DISPUTE_EXPIRED` webhook constants
 
 ## 2.107.0
 * Add `TRANSACTION_REFUND_AUTH_HARD_DECLINED` and `TRANSACTION_REFUND_AUTH_SOFT_DECLINED` validation errors